check exit status of each command

+++ setup.sh
@@ -45,14 +45,23 @@ fi
 
 if [ ! -d eeros ]; then
 	git clone git://github.com/eeros-project/eeros-framework.git eeros -o upstream -b $eeros_branch
+	if [ $? -ne 0 ]; then
+		exit 2
+	fi
 fi
 
 if [ ! -d eeduro ]; then
 	git clone git://github.com/eeros-project/eeduro-platform.git eeduro -o upstream -b $eeduro_branch
+	if [ $? -ne 0 ]; then
+		exit 3
+	fi
 fi
 
 if [ ! -d linaro-tc ]; then
 	wget -c https://launchpad.net/linaro-toolchain-binaries/trunk/2013.03/+download/gcc-linaro-arm-linux-gnueabihf-4.7-2013.03-20130313_linux.tar.bz2
+	if [ $? -ne 0 ]; then
+		exit 4
+	fi
 	echo extracting...
 	tar xjf gcc-linaro-arm-linux-gnueabihf-4.7-2013.03-20130313_linux.tar.bz2
 	rm gcc-linaro-arm-linux-gnueabihf-4.7-2013.03-20130313_linux.tar.bz2
